+++ forui/test/src/theme/color_scheme_test.dart
@@ -48,8 +48,9 @@ void main() {
           error: Colors.blueAccent,
           errorForeground: Colors.blueGrey,
           border: Colors.lime,
+          hoverLighten: 0.3,
+          hoverDarken: 0.2,
           disabledOpacity: 0.1,
-          enabledHoveredOpacity: 0.2,
         );
 
         expect(copy.brightness, equals(Brightness.dark));
@@ -67,8 +68,9 @@ void main() {
         expect(copy.error, equals(Colors.blueAccent));
         expect(copy.errorForeground, equals(Colors.blueGrey));
         expect(copy.border, equals(Colors.lime));
+        expect(copy.hoverLighten, 0.3);
+        expect(copy.hoverDarken, 0.2);
         expect(copy.disabledOpacity, 0.1);
-        expect(copy.enabledHoveredOpacity, 0.2);
       });
     });
 
@@ -95,7 +97,8 @@ void main() {
           ColorProperty('error', Colors.red),
           ColorProperty('errorForeground', Colors.redAccent),
           ColorProperty('border', Colors.lightBlue),
-          PercentProperty('enabledHoveredOpacity', 0.9),
+          PercentProperty('hoverLighten', 0.075),
+          PercentProperty('hoverDarken', 0.05),
           PercentProperty('disabledOpacity', 0.5),
         ].map((p) => p.toString()),
       );